POC/requirements.txt
flask==2.3.2
transformers==4.30.2
torch==1.13.1
torchvision==0.14.1
opencv-python==4.7.0.72
pytesseract==0.3.10
pandas==1.5.3
numpy==1.24.3
Pillow==9.5.0
spacy==3.5.3
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0.tar.gz